requirements.txt
ccxt>=4.0.0
pandas>=1.5.0
matplotlib>=3.5.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
numpy>=1.23.0
scikit-learn>=1.2.0
requests>=2.28.0
pydantic>=1.10.0
torch>=2.0.0
ta>=0.10.2
scikit-image>=0.20.0
optuna>=3.0.0
scipy>=1.10.0
mplfinance>=0.12.9b7 